Moved `TlsConnection` to `transport` module
